Refactor and add rudimentary authentication
