+++ README.md
@@ -1,2 +1,19 @@
 # Tuner_AI
- 
+
+## Server Configuration
+
+The following environment variables are required
+
+* `OPENAI_API_KEY`
+* `PASSWORD`
+
+## Authentication/Authorization
+A rudimentary authentication scheme is implemented. "Login" generates a cookie containing the password.  Subsequent requests to the website
+will make a comparison against that cookie password.
+
+## Server
+To run the server locally
+
+```bash
+npm run server
+```
