--- client/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Boosted 4.0 TunerAI</title>
-    <link rel="stylesheet" href="style.css">
-  </head>
-  <body>
-    <div id="app">
-      <div id="chat_container"></div>
-
-      <form>
-        <textarea name="prompt" rows="1" cols="1" placeholder="Ask Tuner..."></textarea>
-        <button type="submit"><img src="assets/send.svg" /></button>
-      </form>
-    </div>
-
-    <script type="module" src="script.js"></script>
-  </body>
-</html>
